+++ src/StateMachineServiceProvider.php
@@ -6,23 +6,17 @@ use Illuminate\Support\ServiceProvider;
 
 class StateMachineServiceProvider extends ServiceProvider
 {
-    /**
-     * Bootstrap the application services.
-     */
-    public function boot()
+    public function boot(): void
     {
         if ($this->app->runningInConsole()) {
             $this->publishes([
-                __DIR__ . '/../config/config.php' => config_path('state-machine.php'),
+                __DIR__.'/../config/config.php' => config_path('state-machine.php'),
             ], 'caner-state-machine-config');
         }
     }
 
-    /**
-     * Register the application services.
-     */
-    public function register()
+    public function register(): void
     {
-        $this->mergeConfigFrom(__DIR__ . '/../config/config.php', 'state-machine');
+        $this->mergeConfigFrom(__DIR__.'/../config/config.php', 'state-machine');
     }
 }